PlayerStore refactored; functional of track progress moved to PlayerTrackProgress component and TEMP don't work; Info of current playing track moved to PlayerTrackInfo component; Don't tested functional of VolumeSlider created in store
